Use BigIntegerField for ImportExportResource.size

Closes #3424

See merge request baserow/baserow!3146

+++ backend/src/baserow/core/models.py
@@ -678,7 +678,7 @@ class ImportExportResource(CreatedAndUpdatedOnMixin, models.Model):
             "This is only used in the frontend for uploaded files.",
         ),
     )
-    size = models.PositiveIntegerField(
+    size = models.PositiveBigIntegerField(
         default=0, help_text="The size of the resource in bytes."
     )
     created_by = models.ForeignKey(
